CHANGED baseURL TO AMAZON WEB SERVICE Also just exported the URL in App rather than having a separate file for it.
