+++ frontend/src/App.js
@@ -3,6 +3,9 @@ import React, { useState } from 'react';
 
 import Controls from './Components/Controls';
 
+// export const baseURL = "http://localhost:8080/api/settings";
+export const baseURL = "https://jfzlkq6spl.execute-api.ap-southeast-2.amazonaws.com/api/settings";
+
 function App() {
 
   const [threshold, setThreshold] = useState(null);
