--- frontend/src/Components/URL.js
@@ -1,2 +1,0 @@
-export const baseURL = "http://localhost:8080/api/settings";
-// export const baseURL = "https://jfzlkq6spl.execute-api.ap-southeast-2.amazonaws.com/api/settings";
